Added cart on issuing/Added pagination on stocks table/generation receipt/redirectto stocks after login
