+++ app/Http/Controllers/API/ReportController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\API;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Services\ReportService;
+use Illuminate\Support\Facades\DB;
+use Carbon\Carbon;
 use PDF;
 
 class ReportController extends Controller
@@ -39,12 +41,40 @@ class ReportController extends Controller
     }
 
     public function receipt(Request $request){
-        $data = array(
-            'test' => 'test'
-        );
+        $data = [];
+        $items = [];
+        $total_amount = 0;
+        $transactions = DB::table('sales')
+                        ->leftJoin('items', 'sales.item_id','=','items.id')
+                        ->select('sales.*','items.item_code','items.item_name', 'items.price')
+                        ->where('sales.transaction_no', '=', $request['transaction_no'])
+                        ->get();
 
-        $pdf = PDF::loadView('reports.receipt',compact('data'));
-        // return $pdf->download('firstreport.pdf');
-        return $pdf->stream('receipt.pdf');
+        if(count($transactions) > 0) 
+        {
+            foreach ($transactions as $key => $transaction) 
+            {    
+                $items[] = array(
+                    'transaction_id' => $transaction->id,
+                    'customer_name' => $transaction->fullname,
+                    'item_code' => $transaction->item_code,
+                    'item_name' => $transaction->item_name,
+                    'fund' => $transaction->fund,
+                    'transaction_date' => Carbon::parse($transaction->created_at)->format('F d, Y'),
+                    'amount' => $transaction->amount
+                );
+                $total_amount = floatval($total_amount) + floatval($transaction->amount);
+            }
+            $data['items'] = $items;
+            $data['total_amount'] = $total_amount;
+            $data['transaction_no'] = $request['transaction_no'];
+            //dd($data);
+            $pdf = PDF::loadView('reports.receipt',compact('data'));
+            // return $pdf->download('firstreport.pdf');
+            return $pdf->stream('receipt.pdf');
+        }
+        else {
+            return response()->json(['status' => FALSE, 'message' => 'Transaction receipt could not be found'], 400);
+        }
     }
 }
