+++ app/Http/Controllers/API/SalesController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\Sales\StoreSale;
 use App\Http\Requests\Sales\UpdateSale;
 use App\Model\Sale;
 use App\Model\Item;
+use Webpatser\Uuid\Uuid;
 use PDF;
 
 class SalesController extends Controller
@@ -37,31 +38,39 @@ class SalesController extends Controller
      */
     public function store(StoreSale $request)
     {
-        $item = Item::findOrFail($request['item_id']);
+        $transaction_no = strtoupper(uniqid('TR-'));
 
-        if($item->quantity >= $request['quantity']) 
+        foreach ($request['items'] as $key => $req_item) 
         {
-            $sale = new Sale;
-            $sale->customer_type = $request['customer_type'];
-            $sale->customer_id = $request['customer_id'];
-            $sale->fullname = $request['fullname'];
-            $sale->department = $request['department'];
-            $sale->item_id = $request['item_id'];
-            $sale->quantity = $request['quantity'];
-            $sale->remarks = $request['remarks'];
-            $sale->save();
+            $item = Item::findOrFail($req_item['id']);
 
-            $item->quantity = intval($item->quantity) - intval($request['quantity']);
-            $item->save();
+            if($item->quantity >= $req_item['request_quantity']) 
+            {
+                $sale = new Sale;
+                $sale->transaction_no = $transaction_no;
+                $sale->customer_type = $request['customer_type'];
+                $sale->customer_id = $request['customer_id'];
+                $sale->fullname = $request['fullname'];
+                $sale->fund = $request['fund'];
+                $sale->department = $request['department'];
+                $sale->item_id = $req_item['id'];
+                $sale->quantity = $req_item['request_quantity'];
+                $sale->amount = floatval($item->price) * floatval($req_item['request_quantity']);
+                $sale->remarks = $request['remarks'];
+                $sale->save();
 
-            return [
-                'message' => 'Sales has been saved.',
-                'sales_no' => $sale->id,
-            ];
+                $item->quantity = intval($item->quantity) - intval($req_item['request_quantity']);
+                $item->save();
+            }
+            else {
+                return response()->json(['message' => 'Not enough stock for request quantity. (Stocks left: '.$item->quantity.')'], 422);
+            }
         }
-        else {
-            return response()->json(['message' => 'Not enough stock for request quantity. (Stocks left: '.$item->quantity.')'], 422);
-        }
+        
+        return [
+            'message' => 'Sales has been saved.',
+            'transaction_no' => $transaction_no,
+        ];
     }
 
     /**
