+++ app/Model/Sale.php
@@ -6,5 +6,5 @@ use Illuminate\Database\Eloquent\Model;
 
 class Sale extends Model
 {
-    //
+    
 }
