+++ composer.json
@@ -11,7 +11,8 @@
         "fideloper/proxy": "^4.0",
         "laravel/framework": "5.7.*",
         "laravel/passport": "^7.0",
-        "laravel/tinker": "^1.0"
+        "laravel/tinker": "^1.0",
+        "webpatser/laravel-uuid": "2.*"
     },
     "require-dev": {
         "beyondcode/laravel-dump-server": "^1.0",
