+++ resources/js/components/auth/login.vue
@@ -106,8 +106,11 @@
             fetchUser() {
                 this.$store.dispatch('AuthModule/FETCH_USER')
                     .then(response => {
-                        this.isLoading = false;
-                        this.$router.push('/administrator/stocks');
+                        setTimeout(() => {
+                            this.isLoading = false;
+                            // this.$router.push('/administrator/stocks');
+                            window.location = window.location.protocol + "//" + window.location.host + "/administrator/stocks";
+                        }, 1000);
                     })
             }
         }
