+++ resources/js/components/pagination/index.vue
@@ -6,7 +6,7 @@
         
         <div class="col-md-8 float-right">
             <nav aria-label="...">
-                <ul :class="'pagination ' + classSize +' justify-content-end'">
+                <ul :class="'pagination '+' justify-content-end'">
                     <li :class="object.from == object.current_page ? 'page-item disabled' : '' ">
                         <a class="page-link" @click.prevent="firstPage(object.first_page_url)" href="#" tabindex="-1">
                             <span class="fa fa-long-arrow-left"></span>
@@ -17,7 +17,7 @@
                     </li>
                     <template v-for="i in links">
                         <li :class="object.current_page == i ? 'page-item active':'active'"  :key="i">
-                            <a class="page-link" @click.prevent="toPage(i)" href="#">{{ i }}</a>
+                            <a class="page-link" @click.prevent="toPage(i, object)" href="#">{{ i }}</a>
                         </li>
                     </template>
                     <li :class="object.next_page_url == null ? 'page-item disabled' : 'page-item'">
@@ -36,10 +36,8 @@
 
 <script>
     export default {
-        props:['scope','object', 'url', 'classSize', 'query'],
         props: {
             object: {
-                type: Object,
                 required: true
             }
         },
@@ -75,20 +73,21 @@
             }
         },
         methods:{
-            toPage(page){
-                
+            toPage(page, obj){
+                let path = `${obj.path}?page=${page}`; 
+                this.$store.dispatch('ITEMS_MODULE/FETCH_ITEMS', path);
             },
             nextPage (url) {
-               
+                this.$store.dispatch('ITEMS_MODULE/FETCH_ITEMS', url);
             },
             prevPage (url) {
-            
+                this.$store.dispatch('ITEMS_MODULE/FETCH_ITEMS', url);
             },
             firstPage(url){
-                
+                this.$store.dispatch('ITEMS_MODULE/FETCH_ITEMS', url);
             },
             lastPage(url){
-                
+                this.$store.dispatch('ITEMS_MODULE/FETCH_ITEMS', url);
             },
         }
     }
