+++ resources/js/store/modules/cart/index.js
@@ -12,7 +12,7 @@ export default {
             return state.cart_item;
         }
     },
-    mutation: {
+    mutations: {
         CLEAR_ITEM: state => {
             state.cart_item = {};
         },
@@ -20,13 +20,16 @@ export default {
             state.cart_items = [];
         },
         SET_CART_ITEMS: (state, payload) => {
-            state.cart_items = payload;
+            state.cart_items.push(payload);
         },
         SET_CART_ITEM: (state, payload) => {
             state.cart_item = payload;
         },
         REMOVE_CART_ITEM: (state, payload) => {
-            state.cart_items.splice(state.cart_items.indexOf(payload), 1);
+            state.cart_items.splice(payload, 1);
+        },
+        UPDATE_CART: (state, payload) => {
+            state.cart_items[payload.index] = payload.item;
         }
     }
 }
